Drumkit: drop legacy and per-component export
Now only the drumkit as a whole can be exported and just the current version.

Per-component export did work but was always a stub with respect to UX. There was no way to import a component into a kit, fuse several of them or something similar.

Instead of making the user responsible for compatibility with older versions using a legacy export, we drop it and provide forward compatibility and patch releases in the 1.2.X series. This hopefully results in a better user experience and makes our code base more easy to maintain.

Implements MADR 0004