Sampler: include compo gain for pre-fader
it seems during the introduction of `DrumkitComponent` and `InstrumentComponent` both the volume of the former and the gain of the latter were introduced into post fader mixer (in `Sampler` in case per-track output for JACK is used). But the gain of `InstrumentComponent` - part of an `Instrument` and only accessible from within `InstrumentEditor` - was never added to pre fader mixing.

I think it belongs in there as the fader controls the overall volume of an instrument and an instrument component is part of it.