MainForm: remove "Add Component" from menu
it made sense as long as there were `DrumkitComponent`s correlating all `Instrument`s of a kit. But as of now there is `InstrumentComponent` within an `Instrument`. The action present in the Drumkit menu would only act on the currently selected instrument which is not intuitive. Instead, all component handling will be done via `InstrumentEditor`