Instrument: keep track on notes in `enqueue()`
each time an `Instrument` is now enqueued it stores a small string representing the `Note` it was enqueued by. This way we can see a lot better what is going on during debugging.

In addition, the destructor of `Instrument` does now issue a warning in case the instance is still queued.