Fix crash due to undefined order of static initialisers
Filesystem::log_file_path relies on a static-initialised value set
in Filesystem::__usr_log_path. We should avoid calling this from
static initializers because the order of execution of static
initialisers is undefined.

Alternatively (or additionally) removing the dependency on static
initializers in Filesystem::log_file_path would be benificial but
ordering dependencies there are trickier