rework `Drumkit` switching & sample un/loading
For a more smooth transition when switching kits, all associated samples of the old kit are no longer unloaded right away. Instead, all its instruments are added to the death row and `Note`s in our queues can still be rendered.

In addition, `CoreActionController::setDrumkit` does not create a copy of the provided kit anymore. This was moved into the calling routines in the GUI. As a result only the switching itself is undone/redone. All changes in parameter values persist.