rework `Instrument` adding/deletion undo/redo
Previously, whenever an `Instrument` was added using drag&drop or a new instrument was created from scratch the actual `Instrument` creation did happen in the redo part of the action. As a result, hitting Ctrl+Z and Ctrl+Shift+Z would discard all changes applied to the instrument.

I have to admit this is of way less relevance than when changing drumkits but information loss is not nice and consistency in terms of UX is always a good thing. So, instrument adding undo/redo does now store pointers to the `Instrument`s and changes are not discarded. Like it is done for `Drumkit`s.