release during `Drumkit`/`Instrument` switching
instead of letting all samples associated with the old `Instrument`/`Drumkit` ring till the end and processing associated notes still residing in the midi or song note queue, we now jump to the release ADSR phase for those already playing and discard all others. This gives a much smoother transition as soon as longer samples are involved.

Implements #1903