data/xsd: change note position into float
This will not break anything within Hydrogen (but maybe cause problems when positions are edited by hand in XML files).

For v2.0 we touched this file anyway. So, now seems a good time to lay the basis of a potential tuplet support. I have no plans implementing it at the moment. But when one of us goes for it, she could do so without break pattern validity.