Drumkit: harden `install` against missing UTF-8 support
in case the system does not support UTF-8 encoding but either `.h2drumkit` file to be installed or any of the contained entries contain non-Latin1 characters, the content will be installed in a path containing wacky characters. Even though the installation itself did worked, the calling routine will not be able to access the kit as it expects it to be located in a different location.

To circumvent this problem, we remove all non-Latin1 characters for the extracted file and folder names - if and only if `libarchive` complained about missing UTF-8 support - and make the calling routine aware of the actual location.

When installing a kit via the GUI a popup will indicate at least one of the file paths had to be altered.