CoreActinController: harden `retrieveDrumkit`
the temporary folder used to extract a kit for validation checking does not use the name of the kit anymore. This only causes trouble in the presence of non-Latin1 characters in the kits name.

But it turns out that even some characters within the Latin1 range make unit tests fail locally on my Windows machine (with encoding not properly set up). We really have to remove almost everything except of Roman alphabet, Arabic numbers and 1-2 symbols for this to work on such systems