Drumkit: restrict UTF-8 export to Windows with `libarchive` >= 3.5
`libarchive` is quite a pain. Our build pipeline catched an export failure on Ubuntu 20.04 - which I was able to reproduce locally. Well, at least a bit. The problem seems to be this one https://github.com/ libarchive/libarchive/pull/1389/commits/c30f279475e2afd39f380622d2b53b157eb746d8 but it is occurring with `libarchive` installed from Canonicals package repos. Building the same library with the same options and patches (https://launchpad.net/ubuntu/focal/amd64/libarchive13/3.4.0-2ubuntu1) I was not able to reproduce it.

But the result of this bug is quite severe. Not just UTF-8 support is affected Hydrogen is not possible anymore to export drumkits at all.

Since the original issue (#1957) only affects systems having their locale set up inproperly and it was only reported to appear on Windows, I will be a little more conservative in here. The UTF-8 export patch is now only affecting Windows systems with `libarchive` >= 3.5.0. The latter should be guarantueed as we ship the corresponding DLL ourselves.