Filesystem: do not remove common special characters
when cleaning file names on systems with missing UTF-8 support. Else e.g. `Kick - 1.wav` will be moved to `Kick1.wav` and sample could not be loaded anymore. This happens to a lot of our hosted kits.

I also tweaked the kit in our test artifacts to contain whitespaces, [, ], (, ), &, + and -