Filesystem: less supported formats
the original list of supported formats was created based on the `libsndfile` documentation. But I removed some entries from this list which either are not supported by `ffmpeg` and can not be tested by me locally (`mp1`, `l16`) or for which import did not work (`mp2`, `pcm`).

All format conversion have been done using ffmpeg version 5.1.6-0+deb12u1.

`mp3` is a special case and should only be supported for `libsndfile` versions bigger than 1.1. But since the project only provides a version string we always add it. It is quite unlikely for a newer Hydrogen version (containing this code) to be linked against such an old `libsndfile` version and the possible harm caused by a parsing error of a version string defined upstream is probably higher.