tests: add SongExportTest
which does attempt to export a short example song in all supported percolations of interpolation mode, audio file format, sample rate, and sample depth.

I know. this is quite over the top and takes too long for the pipeline. But I want to run it for all platforms at least once and come up with a reduced version in a latter commit