cmake: check for mp3 and flac/ogg support
instead of just displaying a warning message during compilation, we do now write a variable `H2CORE_HAVE_FLAC_SUPPORT` to indicate to the core whether to include FLAC and OGG/Vorbis support or not. This allows to hide options which would not work in the first place. It could be a disadvantage in case Hydrogen is build with one version of `libsndfile` but used with another one during runtime. But this is both an edge case and an anti pattern.

In addition, `H2CORE_HAVE_MP3_SUPPORT` was added to indicate a `libsndfile` version >= 1.1.0. This is key because required symbols for MP3 support would make earlier versions crash.

I also added `H2CORE_HAVE_OPUS_SUPPORT` to indicate whether OGG/Opus in addition to OGG/Vorbis is supported. Not quite sure yet whether we provide that option. But it would be nice.