ExportSongDialog: overhaul
to support all available file formats and expose the compression level for FLAC and formats based on lossy compression.

Instead of WAV, FLAC is now the new default format selected on first opening.

The different WAV and AIFF options in the template combo box were dropped and replaced by a format combo box exposing only the available formats. The previous setup was outdated. In case someone really want to export to WAV that persons most probably do not want to do so with reduced sample rate and depth. For more concise files FLAC or lossy compression will be used.