PatternEditor: add resilience against Note without Instr.
starting from v2.0 Hydrogen will learn to handle `Note`s within a `Pattern` not assigend to any `Instrument` of the current `Drumkit`. Since 1.2.4 is intended to provide forward compatibility and to serve as a rollback for users not satisfied with >=v2.0 we have to harden against those notes.

During final testing for the 1.2.4 release I encountered a segfault on Windows when attempting to open an autosave file created with v2.0. Not quite sure whether this already does the trick, but it will sure help and avoids floods of false positive error log messages