rename JACK Timebase states
while updating the manual for 1.2.4 I decided to drop the term "master" in context of JACK Timebase as well. I did wrote this one way to often for this patch release. The master strip in the mixer is ok (IMO) but the whole master-slave terminology for a more or less recent protocol? Nope.

Language has a huge effects on people and we do not want to reproduce harmful terms in our documentation. Therefore, we will use Timebase "controller" and "listener".

The only places we keep the term "master" is in the `hydrogen.conf` file and the OSC API.

Most prominently the "J.Master" button was renamed to "Timebase".