base `jackTimebaseStateChanged` on event value
instead on relying on the current state found in the `JackAudioDriver` instance.

This is done to allow the shotlist of our documentation to take shots of all possible states of the Timebase button - including the red one.

In addition, it fixes the initial state of the button. When being timebase controller while quitting, Hydrogen remembers and register itself during the next startup. Previously, this was not reflected by the GUI