Timeline: update default BPM
within the `Timeline` we have to so called "special tempo marker": in case no tempo marker is set in the first column, the overall bpm of the current song (the one set via the BPM widget in `PlayerControl`) is used at the beginning of the song till the encounter of the first actual tempo marker.

The value for the bpm of the `Song` is cached in `Timeline` and was only updated when activating the `Timeline` again. This lead to inconsistencies when not having a tempo marker at first column but at a later one, switching to Pattern mode, adjusting speed, and changing back to Song mode.