MainForm: fix redo after shortcut modification
since changing the Preferences triggeres `MainForm::createMenuBar` - to update the shown shortcuts - we have to be very careful to enabling/disabling the proper undo/redo actions based on the current stack position.

Also, it is important to only connect these once. Else the property will be set over and over again.