AudioEngine: do not `reset()` `m_fSongSizeInTicks`
every time the `reset()` function was called in `AudioEngineTests` the song size had to be reapplied.

I think conceptionally it is more elegant _not_ reset the song size in `reset()`. This way it does not reset to its initial state after startup but just resets its transport state. Association with the current song and Hydrogen session still remains