AudioEngine: omit offsets on JACK-based relocation
Within Hydrogen we have to ensure both frame and tick position are consistent during playback. This is important to avoid glitches in audio rendering.

Whenever a tempo change is encountered, the transport position in ticks stays the same but tick size and thus the position in frames changes. To compensate this effect, we introduce a variety of offsets. One for tempo changes, one for song size changes, and another for JACK Timebase stuff.

But these offsets are for internal use only. When relocating via the JACK server, we have to provide `TransportPosition::m_nFrame` without any offsets. When encountering a relocation in `JackAudioDriver::updateTransportPosition` `AudioEngine::locateToFrame` or `AudioEngine::locate` is triggered which resets all existing offsets anyway