PatternEditorPanel: add DB for Pattern editor content
With the introduction of instrument types `PatternEditorInstrumentList` and the remainder of the pattern editor are now decoupled. While the former only lists instruments found in the current drumkit, the latter focus on the notes contained in the current pattern.

We keep established way of listing all instruments of the drumkit in the order defined by the author. But at the bottom we also add additional rows for all the notes not covered by the current drumkit. E.g. You create a note for the second snare and another one for the third of a drumkit and switch to GMRockKit. There you will only find a single snare and the notes are somewhat floating and can not be played back. But they will be still accessible (at a later point) at the buttom of the `DrumPatternEditor` and can be copy/dragged etc. by the user as regular notes. When switching back to the former kit, they can be heard again (without any loss of information).

To allow for this new complexity - and to more easily implemented requested features like collapsing rows in the future - I decided to make `PatternEditorPanel` the single source of truth for which row is associated to which instrument id and type. `PatternEditorPanel::m_db` and `PatternEditorPanel::m_nSelectedRow` have to be used by all other parts of the pattern editor (in commits which will follow)