PatternEditor: rework event handling
for our internal `EventListener`.

Looks like things grew very naturally in here. For almost all (relevant) events `PatternEditorPanel` was calling `updateEditors()` triggering updates of all individual parts of the pattern editor area (except for the `PatternEditorInstrumentList`, which is special). But apart from this most editors did listen to the event as well and were performing a second, redundant update. This is both inefficient and slows down Hydrogen.

Now, as the underlying data for all editors in the pattern editor parts is centralized in `PatternEditorPanel`, all listening and updating will be done from this central spot. Only this way we can ensure the editors work on valid data.