Pattern: use shared_ptr
instead of raw ones.

Sicne the GUI caches a pointer to the currently selected pattern, we get into trouble when this pattern is either deleted or another song is loaded. The GUI still tries to access the already deleted pattern and it takes a couple of micro seconds till a notification event from the core arrives telling the GUI to update its pattern.

Previously, this problem was solved by refetching the selected pattern every time an important rendering/update was done. But this defies the caching and is not that efficient. Plus, it was not done in all places.

Instead, we now use shared pointer for `Pattern` without any explicit delete. This allows the cached pattern in the GUI to stay alive a little longer till the latter is able to get the new one.