PatternEditors: base row on DB
instead of the currently selected instrument.

This allows to handle and select rows not associated to an instrument of the current drumkit. Note that `Note` selection was not ported to use the DB yet.

Selecting a row in the pattern editor and selecting an instrument has been separated. Else, whenever a row without an instrument ID is clicked, incoming MIDI messages would not be played back anymore (since no instrument was selected). This would make pattern editor UX more easy and clear but cause quite some trouble for the MIDI folks. To ease instrument handling in the GUI, `PatternEditorPanel` has now a helper function `getSelectedInstrument` returning the instrument selected in the pattern editor (not core)