Note: move line-pitch conversion into core
in order to allow to properly test it with our existing (core-only) unit tests.

In addition, a couple of magic numbers have been replaced in `PianoRollEditor` and `OCTAVE_NUMBER` was introduced. For new constants I would rather use `const expr` but it would have been inconsistent with respect to the other definitions in `Note.h`