undo/redo instrument renaming
the whole `Instrument` will be replaced. This is quite a waste of resources. But, on the other hand, it is probably done only very rarely. So, it probably does not amount to much in an average undo stack. Thus, I would prefer to keep the code base small and reuse existing routines until this one turns out to be an actual issue