find notes through instrument id & type
pattern editors do now reference and find `Note`s based on `PatternEditorPanel::m_db` and the instrument ids and types it contains instead of `std::shared_ptr<Instrument>` corresponding to a particular row. This way they now should be able to fully handle notes within `Pattern`s not corresponding to any instrument of the current kit (but holding the instrument type of the one they were created with).

To make this happen, `Pattern::findNote` had to be changed to handling instrument ids and types. The underlying promise is that an instrument id always have to be unique. But this should be no problem.

Notes "outside" of the current kit are not rendered yet