Hydrogen: drop `m_nInstrumentLookupTable`
not quite sure why it was introduced in Hydrogen 0.9.6 or what is was intended to do in the future. But since then it was just mapping an index of an array to that very number and is entirely redundant. Moreover it was also dangerous because a C-style array was used without surrounding bound checks