cmake: check for empty libsndfile version
for some reason the variable `PC_LIBSNDFILE_VERSION` in our AppVeyor build pipeline is defined but empty. On the local mac I used for testing `libsndfile` was not found using `pkg-config` causing the variable to be not defined at all. In the remote version, the log suggests it is not found by `pkg-config` either. But the `cmake` version used in there falls back to a different behavior causing the variable to be defined but empty.

Checking for an empty string will hopefully work for all platforms