PatternEditors: right click-drag selection
previously right-click dragging in `DrumPatternEditor` and `PianoRollEditor` did only alter the property shown in `NotePropertiesRuler` and the length of the very note clicked. Now, whenever a note within the current selection is clicked, the whole selection is edited. If a note outside of the selection is dragged, just that single note is edited.