PatternEditors: fix NoteOff handling
for NoteOff notes it makes sense to edit both lead/lag and probability. But velocity, pan, and key/octave not so much.

The former two properties can now be adjusted in `NotePropertiesRuler` and via right click-drag in `DrumPatternEditor` and `PianoRollEditor`. For the later properties, the corresponding note is not rendered in `NotePropertiesRuler`