fix const char handling
when creating a `char*` or `const char*` from a `QString` it is important to _not_ do so with chained commands, like `sString.toUtf8().constData()`. The `QByteArray` returned by `sString.toUtf8()` is transient and may be destructed while its pointer is still in use.

Those statements used in the context of logging I did not touched. In there the data is used right away and I would not expect the behavior above to cause problems in there