PatternEditorSidebar: only allow dragging of instrument lines
for now we create `PatternEditorPanel::m_db` anew every time the underlying drumkit changes and repaint all editors. Moving an instrument to another position in the drumkit is done in the core part of Hydrogen and changes do propagate into GUI.

Changing the position of an instrument type only row - those not corresponding to an instrument of the current drumkit - would be a lot more complex since currently all superfluous types are just ordered alphabetically.

Instead, drag moving is now restrict to rows corresponding to instruments. The instrument type only ones are displayed for convenience and will be made visually different in future commits