PatternEditorPanel: rework instrument selection
I thought about using a separate, dedicated color to indicate the selected instrument in order to better represent the selected instrument in contrast to the selected row in the pattern editor. But this did not look well and made the whole thing even more complex.

Instead, from now on when clicking a type-only column, no instrument will be selected anymore. If in "instrument mode", notes can not be recorded anymore and `InstrumentEditor` will be disabled. This makes instrument handling way more clear