PatternEditors: make note undo actions
independent on `PatternEditorPanel::m_db`.

In its current design we discard and update the DB of the pattern editor each time we switch to a different pattern. As type-only rows may just exist in specific ones, basing the undo/redo note actions on a specific DB row determined by its integer note number will not cut if. Instead, we now provide the content of the row: instrument ID and type to allow finding the note independent of the current state of the DB