NotePropertiesRuler: base dragging on real column
it felt quite strange to only allow changing note properties for those align with the grid in its current resolution. After all, right-click property dragging in `DrumPatternEditor` is possible for all notes. Regardless of position relative to the grid.

With this patch cursor placement and cursor-based property editing is still restricted to the grid, but mouse event apply to all notes