PatternEditor: make single note selection transient
when click dragging a single note not yet part of the selection, selecting it on mouse pressed proved to be harmful. The note(s) must only be selected on mouse movement or click handling will get a lot more complex. In addition, it feels awkward for the note to still be selected once mouse button was released. Its selection is now transient