PatternEditors: fix cursor visibility
if set in `Preferences` the cursor is now hidden on either mouse _press_ or mouse wheel event.

In addition, the cursor is always shown in all editors. Not just the one holding focus. (Focus is indicated differently). In case the causes trouble distinguishing focus later one, we could e.g. set an alpha value to the cursor color in the editor without focus. But having the cursor in all of them makes the whole thing feel much more consistent