PatternEditor: make property drag transient
right-click property drag in `DrumPatternEditor` and `PianoRollEditor` does now behave like left-click drag moving in terms of selection. If the dragged note is part of a selection, the whole selection is altered. If not, the current selection is discarded and the note at point will be assigned to a transient selection for the course of the action