NotePropertiesRuler: fix mouse wheel adjustment
`mouseWheelEvent` does behave like left-click interaction in terms of selection now. If there is a point within the current selection at point, the properties of the whole selection will be adjusted. If there is a point not in the selection, the selection will be discarded and a transient one is created for the course of the action.

In addition, there is no limitation of the row number of notes anymore. When selecting various notes in `DrumPatternEditor`, all of them will show up in `NotePropertiesRuler` and all of them can be adjusted using mouse wheel

The cursor column will not be update on purpose since wheel event can be applied on all notes, regardless of whether they are located on the grid or not