PatternEditor: only update cursor column
when clicked on a grid cell. Not when clicked on a note, which could be off the grid. Having the cursor appear off the position the user clicked at moving the cursor on just a subset of note clickes both felt bad. As a result, clicks in `NotePropertiesRuler` no longer update cursor position