NotePropertiesRuler: rework key press event
`keyPressEvent` works now similar to the other events in terms of selection. In case the cursor is placed on a note within the selection, the action applies to the whole selection. If the note is outside, the selection is discarded and the note at cursor is temporarily selected for the course of the action