NotePropertiesRuler: drop last set value
this was a relict from times before `Selection` was introduced. Using <kbd>Return</kbd> or middle mouse button the value previously assigned to a note could be set to another note as well.

But nowadays `Selection` offers a way more convenient way to set multiple note properties to the same value. In addition, `Selection` swallows the Return key in a very natural way. But before assigning this somewhat hidden feature to another key or dropping keyboard support, we drop it altogether.