PatternEditors: use a margin on mouse cursor interaction
whenever a note is now access by hovering, clicking, or dragging, we do not just select it based on the coordinates of the current mouse cursor but also take a margin into account. When a note is within this margin, it will be selected too. But only the closest one will be picked and if two are of the same distance (left and right of the cursor) the left one will be used.

The margin is dynamic and depends on the current grid resolution. For all resolutions smaller than 1/32 it is set in such a way that allow to keep notes at a 1/32 resolution apart (important when having such a note close to a grid cell and trying to add a new note). When chosing 1/32 resolution, the margin is reduced by half and allows to keep 1/64 notes apart. Beyond that it is `0` and every note can be selected.