PatternEditor: no note deletion on grid click
notes can now only be deleted by clicking them. Previously, clicking the grid cell toggled the presence of a note. But this feels weird now since this only works on a subset of notes and with the cursor margin in place it is as easy as before to delete them