PatternEditor: fix handling of superimposed notes
in case two notes are located at the same location - e.g. one note is moved to the grid position of another and the user cancelled the deduplication dialog - Hydrogen did always delete the first note when moving the selected one away. But is a lot of cases this was wrong, resulting in properties of one of the notes being lost