use shared_ptr for Note
the new PatternEditor patches made the overall application more volatile due to the usage of `PatternEditor::m_selection`. This shared storage of notes contained points to notes which lifecycle was maintained by the audio engine, pattern editors, and `Pattern` class. But whenever a bug led to a note not being properly removed from `m_selection`, we got a use-after-free segfault.

All these selection bug still have to be fixed. But the application should never segfault. In addition, refactoring all basic core classes is on the todo list anyway