PatternEditors: trigger status message for mult. notes
when either a selection is chosen or mutliple notes reside at the same position, the status message shown in Hydrogen was off. A separate message was triggered for each note, they were merge, and only the last one survived, indicating only a subset of the changes applied.

Code was rewritten to compose status messages based on batches of notes. But since selections can contain quite a number of notes quickly, we only trigger status messages for those notes under mouse/keyboard cursor when interacting with selections.