NotePropertiesRuler: rework wheelEvent
instead of allowing to trigger a wheel action on the current selection in all of the `NotePropertiesRuler`, the action does now only take effect when hovering a note. This is done in order to make the UX more consistent as - apart from lasso selection and drawing - all actions do act on hovered notes now.