NoteProperierRuler: sort notes by pitch
in case several notes are located at the same position and will be drawn with a certain X offset, they are no longer order by their insertion time into the corresponding pattern. Instead, they are ordered by pitch. The bottom-most note in `PianoRollEditor` will be rendered first.

This makes it a lot more easy to distinguish the notes shown in the ruler