PatternEditor: sort notes based on pitch
in `DrumPatternEditor` in case there are multiple notes at the same location, the one with the lowest pitch will be rendered now. If there is a selection, the lowest pitched one in the selection is used. Previously, insertion time into the pattern determined which note was shown. But from user perspective this resulted in seemlingly random behavior