PatternEditors: introduce selection/hover border
selection and hover colors are really hard to pick. We use several dim levels depending on whether the note is selected, hovered, or selected and hovered. With all those different colors in `DrumPatternEditor` and `PianoRollEditor` it is really easy to end up with a combination in where the highlight is almost indistinguishable from the background.

To ease this problem, a subtle (mostly transparent) black border was introduced to the selection highlight.