PatternEditorPanel/DrumPatternRow: add bMappedToDrumkit
as a future identifier whether the notes in the row are associated to an instrument within the current drumkit or not.

We need to display _all_ notes. Regardless of whether they have an instrument type assigned or not. This will serve as a distinction between mapped and unmapped notes without type