PatternEditorSidebar: use fixed width for instrument label
in case no instrument is present in one row, this one is _not_ related to the drumkit. However, is this is far from evident in the current design. A first step to change this is to avoid expanding the "plug sign version" of the instrument label into the space usually taken by the mute and solo button