PatternEditors: rename PatternEditor::Mode -> PatternEditor::Property
at first it was solely used to identify which note property ruler was currently shown. But since then it evolved into describing which property of a note is getting edited. And since there are properties the user can edit apart from those in the rulers, a renaming make the code more clear.