EventQueue: introduce Event::Trigger
using this enum class we can clean up event handling within the core.

As an example: switching from pattern mode to song mode does not only trigger EVENT_SONG_MODE_ACTIVATION but also relocation, and selected pattern changed. Loading another song is event worse. It is of course okay to have several events at once in case this is intended. But in most places it did just grew naturally.

Within the GUI this is a problem because we can have race conditions depending on which event arrives first and which side effects of the handling routine is applied first.