PatternEditors: move cursor to clicked note
it feels totally weird when interacting the pattern editor using mouse while keyboard cursor is shown and the latter does not update its position. The note hovered by the keyboard cursor just keeps being displayed all the time. Not good.

This commit moves the keyboard cursor to the hovered /resulting notes most mouse interactions. It intentionally does _not_ move the cursor when 1) opening right-click popup menu without hovering a note and 2) on right-click property drawing in `NotePropertiesRuler`. All other event should update it.

Moving a note with quantization and either a large zoom or big resolution does not work properly yet (as the moved not does not need to be hovered). This will be fixed in an upcoming commit