PatternEditor: add AddNoteAction enum
instead of having the private member `m_bSelectNewNotes`, which arms the pattern editor instance to select newly added notes, a new enum `AddNoteAction` is now passed to `addOrRemoveNoteAction`. Using this one can explicitly request for a specific note to be selected. This is both more clear and serves as a better basis to extend it to more fine grained actions.

Those action will only be applied the first time the note is added. Not on undo/redo