PatternEditor: fix transient selection for popup menu
the previous approach was too dirty and resulted in a bad UX. Everytime a popup was cancelled, the transient selection persists.

Instead, we now use a more sophisticated approach keeping track of both the notes hovered when showing the popup using right click as well as the notes that would be added to the transient selection. In case one of the actions is triggered, the action itself will take care and establishing and clearing the transient selection