NotePropertiesRuler: proper labels for key section
instead of rendering the corresponding note (pitch) names on every background update, we now use proper sidebar labels similar to those available in the `PianoRollEditor`. But there is no popup menu in here.

Next to more efficient rendering the biggest advantage in here is that it is way more easy now to correlate a label with the corresponding column.

In addition, the gap between the sidebar margin and the first note is now properly filled/rendered as well