Sampler: fix mute group rendering
whenever multiple notes at the same position but associate with different instruments of the same mute group were triggered, Hydrogen did render the note that came last (from the perspective of the `Sampler`). For the user, however, this seemed like more or less random behavior.

With this patch we will trigger the note of the bottom-most instrument instead. This way the user knows which note will be rendered and is able to adjust it by rearranging the order of instruments in the current kit