PatternEditor: render effective note length
whenever there is a note of the same mute group to the right of a note, the GUI does now check whether this note might cuase the sample to be stopped prematurely. If so, the effective note length will be rendered using the regular custom length note tail filled with the noteOff color from the preferences.

It is not perfect. It only uses the longest sample, the current tempo, and only the current pattern to check whether the effective length is smaller than the one of the sample. That is done since the pattern could be played back on any position of the song with any speed and an arbitrary pattern following. It is not perfect but is intended to help (especially new) users to understand what is going on when faced with a drumkit with mute groups activated.

In an upcoming commit this rendering option can be toggled in the Preferences