SampleEditor: use whole length as default loop end
(traditionally) Hydrogen does initialize all loop frames with `0`. But we have to be careful to not allow start or loop frame to move beyond the end. Otherwise, resulting sample length would be calculated wrong and cause Hydrogen to freeze (#2092).

(The code design in here is really bad and deserves a rewrite).