SongEditors: rework event handling
(Hydrogen core) event handling in all parts of the song editor was rewritten and aligned to how it is done in the pattern editor.

The individual components, like `SongEditor` or `SongEditorPositionRuler`, are not inheriting from `EventListener` anymore and do not handle events send by the core part of Hydrogen directly anymore. Instead, `SongEditorPanel` listens to all relevant events and triggers the corresponding actions in the individual parts of the song editor. This has the big advantage, that we do not have to care about in event handling anymore. E.g. an event sets a state within its handler in `SongEditorPanel` and the handler in `SongEditor` tries to access it. But there is no predefined order in which handlers for the same event will be triggered.

To ease handling of the overall codebase, the song editor now uses the pattern of `SongEditorPanel::updateEditors()` and `updateEditor()` methods for all individual parts already used in the pattern editor as well.

The overall event handling in `SongEditorPanel` was revised too. Instead of having different event handlers calling each other, the side effects previously performed by the handler were moved to dedicated methods. This way each handler can be maintained independently and the overall design of the song editor is less complex