Mixer: update event handling
previously all batch updates via `Mixer::updateMixer` to faders and rotary also emitted a `valueChanged` signal causing the corresponding instrument strip being selected. But that very behavior is only intended to be used for user interaction.

As a consequence loading a drumkit and applying the new values for faders and rotaries resulted in a lot of additional `EVENT_SELECTED_INSTRUMENT_CHANGED` events causing unnecessary load.

Instead, the `Event::Trigger` enum class is now used in the `Mixer` and its lines too (similar to how it is done in the core part).

A drumkit load does now result in a single, clean `EVENT_DRUMKIT_LOADED` event