Mixer: split line and peak updates
previously the whole `Mixer` was updated with 30 frames per second. Not just the peaks, but all properties of the mixer lines were checked and even whether new ones have to be created or old ones have to be dropped.

Now, peak rendering is split off into `Mixer::updatePeaks` and only this part is trigger by a timer. The update of all properties remains in `updateMixer` for now and it not properly wired yet. This will be ported to event-base updates in following commits.

In addition, C-style arrays for the available `MixerLine` and `LadspaFXMixerLine` have been changed to `std::vector` and some refactoring was done