MixerLine: trigger all actions within widget
instead of emitting signals to `Mixer` and letting this widget communicate the changes to the core. At some point the previous design decision might be useful with future plans taken into account. But right now I do not see a need for such complexity.

I also dropped a somewhat hidden feature to show the InstrumentEditor when double clicking the name of an instrument in the mixer. That is a little bit too over engineered